Fix rustfmt and clippy
